+++ packages/typir-langium/src/features/langium-validation.ts
@@ -5,7 +5,7 @@
  ******************************************************************************/
 
 import { LangiumDefaultCoreServices, Properties, ValidationAcceptor, ValidationChecks } from 'langium';
-import { DefaultValidationCollector, TypirServices, ValidationCollector, ValidationProblem, ValidationRule } from 'typir';
+import { DefaultValidationCollector, LanguageKey, TypirServices, ValidationCollector, ValidationProblem, ValidationRule } from 'typir';
 import { TypirLangiumServices, TypirLangiumSpecifics } from '../typir-langium.js';
 
 export function registerTypirValidationChecks<Specifics extends TypirLangiumSpecifics>(langiumServices: LangiumDefaultCoreServices, typirServices: TypirLangiumServices<Specifics>) {
@@ -86,7 +86,7 @@ export class DefaultLangiumTypirValidator<Specifics extends TypirLangiumSpecific
             accept(problem.severity, message, {
                 // these properties are named differently in Langium and Typir:
                 node: problem.languageNode,
-                property: problem.languageProperty as Properties<Specifics['LanguageType']>,
+                property: problem.languageProperty as (Properties<Specifics['LanguageType']> | undefined),
                 index: problem.languageIndex,
                 // copy all other DiagnosticInfo properties:
                 ...problem,
@@ -110,7 +110,7 @@ export class DefaultLangiumTypirValidator<Specifics extends TypirLangiumSpecific
  * @param T a type definition mapping language specific type names (keys) to the corresponding types (values)
  */
 export type LangiumValidationRules<Specifics extends TypirLangiumSpecifics> = {
-    [K in keyof Specifics['AstTypes']]?: Specifics['AstTypes'][K] extends Specifics['LanguageType'] ? ValidationRule<Specifics, Specifics['AstTypes'][K]> | Array<ValidationRule<Specifics, Specifics['AstTypes'][K]>> : never
+    [K in LanguageKey<Specifics>]?: Specifics['LanguageKeys'][K] extends Specifics['LanguageType'] ? ValidationRule<Specifics, Specifics['LanguageKeys'][K]> | Array<ValidationRule<Specifics, Specifics['LanguageKeys'][K]>> : never
 } & {
     AstNode?: ValidationRule<Specifics, Specifics['LanguageType']> | Array<ValidationRule<Specifics, Specifics['LanguageType']>>;
 }